% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plot.R
\name{change_ylabs}
\alias{change_ylabs}
\title{Change y tick marks of bayesplot}
\usage{
change_ylabs(mcmc_plot, ..., labs = character(), top_down = TRUE)
}
\description{
Misc. function to change y tick marks of bayesplot
}
