% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapeated_kfold.R
\name{extract_K_fold}
\alias{extract_K_fold}
\title{Extract hold-out individuals' parameters from K Fold}
\usage{
extract_K_fold(
  list_of_stanfits,
  list_of_holdouts,
  pars = NULL,
  ...,
  holdout = TRUE
)
}
\value{
a list of array of extracted values
}
\description{
Function to extract hold-out individual parameters aka parameters that have the same dim as number of observation
}
