% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapeated_kfold.R
\name{extract_log_lik_K}
\alias{extract_log_lik_K}
\title{Extract loglik from K-Fold CV
Copied from AhVehtari}
\usage{
extract_log_lik_K(
  list_of_stanfits,
  list_of_holdout,
  loglik_par = "log_lik",
  ...
)
}
\arguments{
\item{list_of_stanfits}{list of stanfits, output from stan_kfold}

\item{list_of_holdout}{a list of holdoult, output from repeated_kfold$holdout}
}
\description{
This function extract loglik from repeated K Fold CV, modified from code of Stan dev
}
