% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapeated_kfold.R
\name{extract_variables}
\alias{extract_variables}
\title{#xtract variables from a stanfit of KFold stanfit}
\usage{
extract_variables(x, variables, ..., merge_chains = FALSE)
}
\arguments{
\item{x}{a draw array}

\item{variables}{a vector of variables}

\item{merge_chains}{merge all chains to one}
}
\value{
an array of 3 dimensions: iteration, chain, and variables if merge = FALSE or a matrix of 2 dimensions: draw and variable if merge = TRUE

a draw matrix
}
\description{
Function to extract variables from draws array
}
