% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapeated_kfold.R
\name{kfold.kfoldll}
\alias{kfold.kfoldll}
\title{Computation of ELPD}
\usage{
\method{kfold}{kfoldll}(log_lik_heldout)
}
\arguments{
\item{log_lik_heldout}{come from extract_log_lik_K}
}
\value{
an object of class c("kfold", "loo"). See \link[loo]{kfold}
}
\description{
method for computation of elpd in loo
}
