% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plot.R
\name{mcmc_intervals_multi}
\alias{mcmc_intervals_multi}
\title{Function to create a side by side mcmc intervals}
\usage{
mcmc_intervals_multi(
  fits = list(),
  multi_point_est = FALSE,
  pars = character(),
  regex_pars = character(),
  transformations = list(),
  ...,
  prob = 0.5,
  prob_outer = 0.9,
  point_est = c("median", "mean", "none"),
  outer_size = 0.5,
  inner_size = 2,
  point_size = 4,
  rhat = numeric()
)
}
\arguments{
\item{fits}{a list of stan fits, must be wrapped in \code{list()}}

\item{multi_point_est}{default to FALSE, whether to set show multiple point est, if TRUE, must specify which point_est you want to plot}

\item{...}{other params passed to \code{bayesplot::mcmc_intervals_data}}
}
\description{
This function is similar to \code{bayesplot::mcmc_intervals} but
provides support for pooling several intervals from several fits.
}
