% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_residual.R
\name{pairwise_corr_plot}
\alias{pairwise_corr_plot}
\title{Parwise correlation plot for LCA}
\usage{
pairwise_corr_plot(obs, pred, B = 0)
}
\arguments{
\item{obs}{a matrix of observed binary outcomes}

\item{pred}{a matrix of predicted probability for corresponding outcome}

\item{B}{number of bootstrap}
}
\description{
Function to calculate pairwise residual correlation
}
\seealso{
(ttps://www.jstor.org/stable/2533043?origin=crossref)
}
