% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapeated_kfold.R
\name{repeated_kfold}
\alias{repeated_kfold}
\title{Split data to K for N times}
\usage{
repeated_kfold(K = 10, N_rep = 4, N_obs, data, seed = 123, cores = 10)
}
\arguments{
\item{K}{number of folds}

\item{N_rep}{number of repetition}

\item{N_obs}{number of observation in the full dataset}

\item{data}{a list that is supposed to feed into Stan}

\item{seed}{seed}

\item{cores}{number of cores to be used. On Windows, will fall back to single core.}
}
\value{
a nested list with 3 elements:
\itemize{
\item keptin: list of keptin(1) or holdout(0) in each folds
\item holdout: negated version of keptin
\item inputs: processed dataset with a keptin variable injected in
}
}
\description{
Create K fold splitted stan inputs N times
}
