% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapeated_kfold.R
\name{stan_kfold}
\alias{stan_kfold}
\title{Stan fit with K fold CV}
\usage{
stan_kfold(
  file,
  sampler,
  list_of_datas,
  include_paths = NULL,
  sample_dir = NULL,
  backend = c("cmdstanr", "rstan"),
  chains,
  cores = getOption("mc.cores", default = 1),
  seed,
  pars = NA,
  merge = TRUE,
  ...
)
}
\arguments{
\item{file}{a stan file}

\item{sampler}{a precompiled stanmodel or CmdStanModel object}

\item{list_of_datas}{list of data, if outputted by repeated_kfold: outputs$inputs}

\item{include_paths}{path to included file. See rstan/cmdstanr help documentation}

\item{sample_dir}{directory to put sample csv files}

\item{backend}{either rstan or cmdstanr. if sampler is available, will be chosen accordingly.}

\item{chains}{number of MCMC chains}

\item{cores}{number of cores to use}

\item{seed}{seed}

\item{pars}{list of parameters to be included}

\item{merge}{whether to merge the samples to one}

\item{...}{other parameters passed to engine call (rstan::sampling or CmdStanModel$sample)}
}
\value{
If merge == TRUE: a stanfit object (if backend = rstan) or sample array (if backend = cmdstanr)
If merge == FALSE: a list of stanfit objects (if backend = rstan) or list of CmdStanMCMC objects (if backend = cmdstanr)
}
\description{
Perfoming Stan fit for K-Fold CV data created by repeated_kfold
}
